requirements.txt
Flask==0.12.0
Flask-Cors==2.1.2
Flask-Migrate==2.0.2
Flask-Script==2.0.3
Flask-SQLAlchemy==2.1
Flask-Security==1.7.5
Flask-Testing==0.6.2
click==6.6
GitPython==2.1.3
PyGithub==1.34
coala
coala-bears
raven[flask]
